+++ app/views/users/sessions_by_code/new.html.slim
@@ -29,8 +29,8 @@
     .fr-mt-8w.fr-mb-4w
       - if @existing_login_code&.very_recent?
         |> Vous pouvez choisir un autre mode de connexion depuis la
-        => link_to "page de connexion", new_user_session_path(login_code: { email: @email }), class: "fr-link"
+        => link_to "page de connexion", new_user_session_path(demande_login_code_form: { email: @email }), class: "fr-link"
       - else
         |> Si vous n’avez pas reçu l’email contenant le code, vous pouvez en demander un nouveau depuis la
-        => link_to "page de connexion", new_user_session_path(login_code: { email: @email }), class: "fr-link"
+        => link_to "page de connexion", new_user_session_path(demande_login_code_form: { email: @email }), class: "fr-link"
         |> Vous pouvez aussi choisir un autre mode de connexion depuis cette page.
